+++ sentry/src/main/java/io/sentry/util/HttpUtils.java
@@ -1,13 +1,22 @@
 package io.sentry.util;
 
+import static io.sentry.util.UrlUtils.SENSITIVE_DATA_SUBSTITUTE;
+
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
+import java.util.Enumeration;
 import java.util.List;
 import java.util.Locale;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 @ApiStatus.Internal
 public final class HttpUtils {
+
+  public static final String COOKIE_HEADER_NAME = "Cookie";
+
   private static final List<String> SENSITIVE_HEADERS =
       Arrays.asList(
           "X-FORWARDED-FOR",
@@ -23,7 +32,102 @@ public final class HttpUtils {
           "X-CSRFTOKEN",
           "X-XSRF-TOKEN");
 
+  private static final List<String> SECURITY_COOKIES =
+      Arrays.asList(
+          "JSESSIONID",
+          "JSESSIONIDSSO",
+          "JSSOSESSIONID",
+          "SESSIONID",
+          "SID",
+          "CSRFTOKEN",
+          "XSRF-TOKEN");
+
   public static boolean containsSensitiveHeader(final @NotNull String header) {
     return SENSITIVE_HEADERS.contains(header.toUpperCase(Locale.ROOT));
+  }
+
+  public static @Nullable List<String> filterOutSecurityCookiesFromHeader(
+      final @Nullable Enumeration<String> headers,
+      final @Nullable String headerName,
+      final @Nullable List<String> additionalCookieNamesToFilter) {
+    if (headers == null) {
+      return null;
+    }
+
+    return filterOutSecurityCookiesFromHeader(
+        Collections.list(headers), headerName, additionalCookieNamesToFilter);
+  }
+
+  public static @Nullable List<String> filterOutSecurityCookiesFromHeader(
+      final @Nullable List<String> headers,
+      final @Nullable String headerName,
+      final @Nullable List<String> additionalCookieNamesToFilter) {
+    if (headers == null) {
+      return null;
+    }
+
+    if (headerName != null && !"Cookie".equalsIgnoreCase(headerName)) {
+      return headers;
+    }
+
+    final @NotNull ArrayList<String> filteredHeaders = new ArrayList<>();
+
+    for (final String header : headers) {
+      filteredHeaders.add(
+          HttpUtils.filterOutSecurityCookies(header, additionalCookieNamesToFilter));
+    }
+
+    return filteredHeaders;
+  }
+
+  public static @Nullable String filterOutSecurityCookies(
+      final @Nullable String cookieString,
+      final @Nullable List<String> additionalCookieNamesToFilter) {
+    if (cookieString == null) {
+      return null;
+    }
+    try {
+      final @NotNull String[] cookies = cookieString.split(";", -1);
+      final @NotNull StringBuilder filteredCookieString = new StringBuilder();
+      boolean isFirst = true;
+
+      for (String cookie : cookies) {
+        if (!isFirst) {
+          filteredCookieString.append(";");
+        }
+
+        final @NotNull String[] cookieParts = cookie.split("=", -1);
+        final @NotNull String cookieName = cookieParts[0];
+        if (isSecurityCookie(cookieName.trim(), additionalCookieNamesToFilter)) {
+          filteredCookieString.append(cookieName + "=" + SENSITIVE_DATA_SUBSTITUTE);
+        } else {
+          filteredCookieString.append(cookie);
+        }
+        isFirst = false;
+      }
+
+      return filteredCookieString.toString();
+    } catch (Throwable t) {
+      return null;
+    }
+  }
+
+  public static boolean isSecurityCookie(
+      final @NotNull String cookieName,
+      final @Nullable List<String> additionalCookieNamesToFilter) {
+    final @NotNull String cookieNameToSearchFor = cookieName.toUpperCase(Locale.ROOT);
+    if (SECURITY_COOKIES.contains(cookieNameToSearchFor)) {
+      return true;
+    }
+
+    if (additionalCookieNamesToFilter != null) {
+      for (String additionalCookieName : additionalCookieNamesToFilter) {
+        if (additionalCookieName.toUpperCase(Locale.ROOT).equals(cookieNameToSearchFor)) {
+          return true;
+        }
+      }
+    }
+
+    return false;
   }
 }